trading/requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-filter==21.1
django-log-viewer==1.1.4
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-api-logger==1.0.10
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
pkg-resources==0.0.0
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
zipp==3.6.0
